api/requirements.txt
passlib[bcrypt]
bcrypt>=4.0.0,<5.0.0
requests
pymongo
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
httpx
beautifulsoup4
lxml
scrapy
crochet

